refactor(query-tools): extract group tools into group-tools.ts, re-export from query-tools
